refactor : rename writeStates
